requirements.txt
langchain
praw
python-dotenv
langchain-groq